@@ -1,23 +1,23 @@
 name: new_version_plus
 description: Check if your user has the most recent version of your Flutter app.
-version: 0.1.0
+version: 0.1.1
 homepage: https://github.com/CodesFirst/new_version_plus
 
 environment:
-  sdk: ^3.6.1
+  sdk: ^3.8.0
   flutter: '>=3.3.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  package_info_plus: ^8.1.3
-  http: ^1.3.0
+  package_info_plus: ^8.3.0
+  http: ^1.4.0
   url_launcher: ^6.3.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
